separate API and implementation
