+++ slf4cats-example/src/main/resources/logback.xml
